Adding a ping timer
This change adds a timer with an interval of half the max ping time. This is used to send Ping messgaes to the server. So far, this is just to make things work: we really want to use the ping OK return messages to ensure the server is still up.