Correcting the KiirooCmd message
The Position field is incorrect and needed replacing with the
Command field. To help with usage, a getter and setter have
been added to aid mapping the numeric position with the string
value.

Also, fixing the build/test configuration so that it works on
Windows as well as *nix.

Fixes #21