fix: Fix noble characteristic notifications
We were using notifications incorrectly. We should subscribe, then
listen for a data event, which has the value attached. Not sure why
this was triggering reads per notification (which was causing the
macOS profiling spikes), but that was very wrong, and was causing the
issues with Lovense detection. Removed specialization code for Lovense
again. Fleshlight Launch still having problems though.

Fixes #180