Update linearAttributes() to use LinearCmd
This seems like a copy-and-paste error.

On attempting to connect to a device with only linear capabilities (in this case, Kiiroo Onyx+) linearAttributes reports nothing, despite LinearCmd in the messages list. 
The linear() command isn't rejected, as LinearCmd is defined, but if following the examples in the dev guide and checking device.linearAttributes.length first, the results will be misleading.